# Improve performance of pulse.Schedule