<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary
Address #2817. This PR improves the run time of the example code given in #2817 from 34s (273s) to 1.5s (6.6s) by improving the implementation of `TimeslotCollection.merged()` and `Schedule.__init__()`.

### Details and comments
The bottleneck was in the reconstruction of `TimeslotCollection` in the constructor of `Schedule`. This PR improves the performance of `Schedule.__init__()` by improving the implementation of `TimeslotCollection.merged()` and using it in `Schedule.__init__()`.