@itoko 2019-07-24
Yes, unless users intentionally enable the option `check_overlap=False`. But even so I have to admit current this PR changes current interface. I understand this PR should not include any interface change and focus only on performance improvement. I'll try to tweak more to improve performance without changing interface.