@itoko 2019-07-24
Now this PR is updated not to allow invalid schedule again, while it still improves performance as shown in the top comment.