@lcapelluto 2019-07-24
I LOVE the type hint fixes. However, please make separate PRs next time 🙏 One PR for type hints, one for "Improve performance of pulse.Schedule". Makes the life of the review much easier