@itoko 2019-07-29
Based on the discussion on this PR, #2870 is going to fix the issue #2817 more completely. Thank you, @taalexander. This PR should be closed if #2870 is merged.