@itoko 2019-08-06
Closed since #2870 is merged.